refactor: downloadNextNftInQueue as action
