+++ packages/desktop/App.svelte
@@ -45,7 +45,7 @@
         resetRouters,
     } from './lib/routers'
     import { openSettings } from './lib/routers/actions/openSettings'
-    import { downloadNextItemInQueue, nftDownloadQueue } from '@core/nfts'
+    import { downloadNextNftInQueue, nftDownloadQueue } from '@core/nfts'
 
     appStage.set(AppStage[process.env.STAGE.toUpperCase()] ?? AppStage.ALPHA)
 
@@ -87,7 +87,7 @@
     $: isDashboardVisible = $appRoute === AppRoute.Dashboard && $hasLoadedAccounts && $popupState.id !== 'busy'
     $: isWindows = $platform === PlatformOption.Windows
 
-    $: $nftDownloadQueue, downloadNextItemInQueue()
+    $: $nftDownloadQueue, downloadNextNftInQueue()
 
     let splash = true
     let settings = false
