+++ packages/shared/lib/core/nfts/actions/index.ts
@@ -1,4 +1,5 @@
 export * from './addOrUpdateNftInAllAccountNfts'
+export * from './downloadNextNftInQueue'
 export * from './getNftByIdFromAllAccountNfts'
 export * from './loadNftsForActiveProfile'
 export * from './setAccountNftsInAllAccountNfts'
